doc: improve example notebook presentation (#16)
<!--
Please use a PR title that conforms to *conventional commits*:
"<commit_type>: Describe your change"; for example: "fix: prevent race
condition". Some other commit types are: fix, feat, ci, doc, refactor...
For a full list of commit types visit
https://www.conventionalcommits.org/en/v1.0.0-beta.2/
-->

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->

Basic demo:

- [X] Added heading structure
- [X] Stated intended learning outcomes
- [X] Included description of function composition with JAX
compatibility

CFD demo:

- [x] Added heading structure
- [x] Stated intended learning outcomes
- [x] Brief explanation of what's going on
- [X] Replaced verbose print output with tqdm
- [X] Included @jpbrodrick89's very cool LBFGS optimiser
- [X] Embedded animation for interactive play in Jupyter Notebook

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

None.

#### License

- [X] By submitting this pull request, I confirm that my contribution is
made under the terms of the [Apache 2.0
license](https://pasteurlabs.github.io/tesseract-jax/LICENSE).
- [X] I sign the Developer Certificate of Origin below by adding my name
and email address to the `Signed-off-by` line.

<details>
<summary><b>Developer Certificate of Origin</b></summary>

```text
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.


Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
```

</details>

Signed-off-by: [YOUR NAME] <[YOUR EMAIL]>

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>