feat: Selective abstract eval (#85)
#### Relevant issue or PR

Tesseract need to have an abstract eval defined, even if jax would not
require them to.

#### Description of changes

This PR avoids calls to abstract_eval if it is not defined in the
Tesseract API. If a function transformation is applied without the
abstract eval, an error is raised.
Furthermore we add a concrete implementation "tesseract_dispatch()" that
does not rely on any function transformations.

#### Testing done

added new tests